Improved mzxml readout
When checking for mass differences for molecular ion identification in raw data, instead of summing up all scans in the retention time range of the deconvoluted peak, now it decides on a single scan based on where the intensity of the ion of interest is the highest. Also, molecular ion candidates below 1% total intensity per scan (not deconvoluted spectrum!) are discarded. This should reduce false positive molecular ion assignments.